[UPD] restful
This commit improves the code for handling the x-www-form-urlencoded content-type and passing parameters in the correct format.
Before this commit, when calling an API with x-www-form-urlencoded and passing parameters, the fields could not adapt well if there were relational fields with record-related fields (e.g., many2one) because the parameters were passed as strings.